requirements.txt
Django==2.2.1
djangorestframework==3.9.4
mysqlclient==1.4.2.post1
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
